refactor(net): split connect, send, and named-unix handlers into submodules
